refactor: move rules folder into src/core
