+++ src/core/rules/index.ts
