+++ tests/rules/country-phone-validator.test.ts
@@ -1,4 +1,4 @@
-import { CountryPhoneValidator } from '../../src/rules/phone/country-phone-validator';
+import { CountryPhoneValidator } from '../../src/core/rules/phone/country-phone-validator';
 
 describe('CountryPhoneValidator', () => {
   it('should validate a phone number in Togo', () => {
